Submit vlog shell script.
1. vlog was written in May 2009 for providing better experiences when
troubleshooting on NetBackup.
As it's not so convinience to troubleshooting on logs from UNIX for
Netbackup.
This shell script is to provide better experiences to view/list/delete
Netbackup logs and shutdown/restart NetBackup processes on UNIX/Linux.
So this is a smart cross-platform shell scriopt.
For details, please refer to the usage and give a try.